+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 This log should always be updated when doing backwards incompatible changes, resulting in a major version bump. Feel free to add a log for lesser version bumps as well, but for major bumps it's a must.
 
+**unreleased**
+-----
+- Changed
+    - Graypy dependency is not installed by default. To install with Graylog support, install with the `graylog` extra: `pip install loggo[graylog]`
+
 6.0.0
 -----
 - Added